requirements.txt
Django==5.1.2
django-ckeditor==6.7.1
pillow==11.0.0